+++ src/java/org/eclim/plugin/jdt/command/src/JavaCommand.java
@@ -72,6 +72,8 @@ import org.eclipse.jdt.core.search.SearchParticipant;
 import org.eclipse.jdt.core.search.SearchPattern;
 import org.eclipse.jdt.core.search.SearchRequestor;
 
+import com.google.gson.Gson;
+
 /**
  * Command to run the project's main class.
  *
@@ -162,6 +164,21 @@ public class JavaCommand
 
     java.setClasspath(classpath);
 
+    // add default vm args
+    String setting = ProjectUtils.getSetting(project, "org.eclim.java.run.jvmargs");
+    if (setting != null && !setting.trim().equals(StringUtils.EMPTY)){
+      String[] defaultArgs = (String[])new Gson().fromJson(setting, String[].class);
+      if (defaultArgs != null && defaultArgs.length > 0){
+        for(String vmarg : defaultArgs){
+          if (!vmarg.startsWith("-")){
+            continue;
+          }
+          Argument a = java.createJvmarg();
+          a.setValue(vmarg);
+        }
+      }
+    }
+
     // add any supplied vm args
     String[] vmargs = commandLine.getValues(VMARGS_OPTION);
     if (vmargs != null && vmargs.length > 0){
@@ -174,7 +191,7 @@ public class JavaCommand
       }
     }
 
-    // add any supplied vm args
+    // add any supplied system properties
     String[] props = commandLine.getValues(SYSPROPS_OPTION);
     if (props != null && props.length > 0){
       for(String prop : props){